+++ src/commands/emit.js
@@ -1,12 +1,18 @@
 "use strict";
 
 const chalk 			= require("chalk");
+const _ 				= require("lodash");
 const { convertArgs } 	= require("../utils");
 
 module.exports = function(vorpal, broker) {
 	// Register broker.emit
 	vorpal
 		.command("emit <eventName>", "Emit an event")
+		.autocomplete({
+			data() {
+				return _.uniq(broker.registry.getEventList({}).map(item => item.event.name));
+			}
+		})
 		.allowUnknownOptions()
 		.action((args, done) => {
 			const payload = convertArgs(args.options);