+++ CHANGELOG.md
@@ -8,13 +8,14 @@
 - new command: `bench [options] <action> [jsonParams]`: Benchmark services
 - new command: `env`: List of environment variables
 - new command: `clear`: Clear cacher
+- add autocomplete to `emit` and `bench` as well
 - extendable with custom commands via ServiceBroker options (`replCommands`)
 - new options in `call` & `dcall` commands:
     - `--load` - load params from a JSON file. The filename generated from action name. E.g. for `posts.find`˙action the filename is `<current_dir>/posts.find.params.json`
-    - `--load filename` - load params from the specified JSON file
+    - `--load [filename]` - load params from the specified JSON file
     - `--save` - save response to file. The filename generated from action name. E.g. for `posts.find`˙action the filename is `<current_dir>/posts.find.response.json`. The extension is `.json` when the response is `object`. Otherwise it is `.txt`.
-    - `--save filename` - save reponse to the specified file
+    - `--save [filename]` - save reponse to the specified file
 - new options in `nodes` command:
     - `--raw` - print entire service registry to then console as JSON
-    - `--file <filename>` - save then entire service registry to the specified file as JSON
+    - `--save [filename]` - save then entire service registry to the specified file as JSON